i3, polybar: back to the gaps hack, spacer doesn't work
with a spacer polybar, the spacer technically renders on top
of of the restacked real polybar, stealing its focus even though
the window is transparent. The end-result is that the real polybar
can't be clicked/interacted with. If I don't restack the real polybar
it will render on top, but then it also renders on top of fullscreen
windows which is undesireable